Move Limelight.kt to lib/
