+++ src/main/java/frc/robot/subsystems/LimelightSubsystem.kt
@@ -1,13 +1,11 @@
 package frc.robot.subsystems
 
-import edu.wpi.first.networktables.NetworkTable
-import edu.wpi.first.networktables.NetworkTableEntry
 import edu.wpi.first.networktables.NetworkTableInstance
-import edu.wpi.first.wpilibj.shuffleboard.Shuffleboard
-import edu.wpi.first.wpilibj.shuffleboard.ShuffleboardTab
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
+import lib.Limelight
 
 object LimelightSubsystem : SubsystemBase() {
+
     private val limelight : Limelight = Limelight(NetworkTableInstance.getDefault().getTable("limelight"))
 
     val xOffset: Double
